Moved the incomplete_registration code to userid_detail model.
1. Removed the unnecessary new class file for incompleted registration
methods.

2. Moved the methods in to userid_detail model.

3. Removed unnecessary breadcrumbs.
