+++ config/breadcrumbs.rb
@@ -336,7 +336,7 @@ crumb :role_listing do
 end
 crumb :incomplete_registrations do
   link 'Incomplete Registration Listing'
-  parent :regmanager_userid_options, incomplete_registrations_userid_details_path
+  parent :root
 end
